fix: correctly handle fallbacks for webpack 5
