+++ test/integration/fallback-modules/fixtures/with-crypto/pages/index.js
